+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/tips/TipsScreen.kt
@@ -4,6 +4,7 @@ import ai.mcpdirect.studio.app.Screen
 import ai.mcpdirect.studio.app.generalViewModel
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -11,9 +12,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.paddingFromBaseline
 import androidx.compose.foundation.layout.width
-import androidx.compose.material3.Button
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -24,11 +23,10 @@ import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
 import mcpdirectstudioapp.composeapp.generated.resources.Res
 import mcpdirectstudioapp.composeapp.generated.resources.keyboard_arrow_right
-import mcpdirectstudioapp.composeapp.generated.resources.mcpdirect_diagram_tips
+import mcpdirectstudioapp.composeapp.generated.resources.mcpdirect_tips_one_url
 import org.jetbrains.compose.resources.painterResource
 
 @Composable
@@ -38,15 +36,10 @@ fun TipsScreen() {
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
         Text(
-            "MCPdirect - Universal MCP Access Gateway",
+            "Let MCP power your business",
             Modifier.padding(vertical = 16.dp),
             style = MaterialTheme.typography.displayMedium)
 //        https://storage.googleapis.com/mcpdirect-image/tips/mcpdirect_diagram_simple.png
-        Image(
-            painterResource(Res.drawable.mcpdirect_diagram_tips),
-            contentDescription = "MCPdirect: Universal MCP Access Gateway",
-            Modifier.width(900.dp),
-        )
         Spacer(Modifier.height(16.dp))
         OutlinedCard(Modifier.width(900.dp),) {
             Row(Modifier.padding(start = 16.dp, end = 4.dp),verticalAlignment = Alignment.CenterVertically) {
@@ -63,16 +56,23 @@ fun TipsScreen() {
             }
 
             HorizontalDivider()
-            Row(Modifier.padding(16.dp).fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween) {
-                TextButton(onClick = {  }) {
-                    Text("1. Connect MCP servers to MCPdirect")
-                }
-                TextButton(onClick = {  }) {
-                    Text("2. Generate MCPdirect key for MCP servers access")
-                }
-                TextButton(onClick = {  }) {
-                    Text("3. Configure MCPdirect in AI Agents")
+            Box{
+                Image(
+                    painterResource(Res.drawable.mcpdirect_tips_one_url),
+                    contentDescription = "MCPdirect: Universal MCP Access Gateway",
+//                    Modifier.padding(top = 8.dp, start = 16.dp, end = 16.dp, bottom = 16.dp).fillMaxWidth().clip(CardDefaults.shape),
+                )
+                Row(Modifier.padding(horizontal = 8.dp).fillMaxWidth(),
+                    horizontalArrangement = Arrangement.SpaceBetween) {
+                    TextButton(onClick = {  }) {
+                        Text("1. Connect MCP servers to MCPdirect")
+                    }
+                    TextButton(onClick = {  }) {
+                        Text("2. Generate MCPdirect key for MCP servers access")
+                    }
+                    TextButton(onClick = {  }) {
+                        Text("3. Configure MCPdirect in AI Agents")
+                    }
                 }
             }
         }